+++ src/main/kotlin/shop/itbug/fluttercheckversionx/util/NotifactionUtil.kt
@@ -1,13 +1,8 @@
 package shop.itbug.fluttercheckversionx.util
 
-import com.intellij.ide.BrowserUtil
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
-import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.Project
-import shop.itbug.fluttercheckversionx.common.MyDumbAwareAction
-import shop.itbug.fluttercheckversionx.document.copyTextToClipboard
-import shop.itbug.fluttercheckversionx.i18n.i18n
 import shop.itbug.fluttercheckversionx.icons.MyIcons
 
 //通知相关工具类
@@ -21,12 +16,12 @@ class MyNotificationUtil {
                 .createNotification(message, type).apply {
                     icon = MyIcons.flutter
                 }
-                .addAction(object : MyDumbAwareAction({ "bug-feedback".i18n() }) {
-                    override fun actionPerformed(e: AnActionEvent) {
-                        message.copyTextToClipboard()
-                        BrowserUtil.open("https://github.com/mdddj/dd_flutter_idea_plugin/issues")
-                    }
-                })
+//                .addAction(object : MyDumbAwareAction({ "bug-feedback".i18n() }) {
+//                    override fun actionPerformed(e: AnActionEvent) {
+//                        message.copyTextToClipboard()
+//                        BrowserUtil.open("https://github.com/mdddj/dd_flutter_idea_plugin/issues")
+//                    }
+//                })
                 .notify(project)
         }