+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## [6.5.0] - 2025-11-04
+
+* Adds provider panel (dart vm) beta version
+* Add `dart pub run build` command run in background action (freezed,foot status bar)
+* Add `open in` popup actions in pubspec.yaml with local package item
+
+## [6.4.3] - 2025-10-31
+
+* Adaptation AS otter version (2025.1)
+
 ## [6.4.2] - 2025-10-30
 
 * fix: media_kit crashes bugs in debug mode ([#95](https://github.com/mdddj/dd_flutter_idea_plugin/issues/95))